Update .NET Framework version to 4.8.
