+++ source/NetFramework/Server/Server.csproj
@@ -52,9 +52,6 @@
   <ItemGroup>
     <Reference Include="System.Configuration" />
     <Reference Include="System.Net.Http" />
-  </ItemGroup>
-
-  <ItemGroup Condition=" '$(TargetFramework)' == 'net48' ">
     <Reference Include="System.Web" />
     <Reference Include="Microsoft.CSharp" />
     <Reference Include="Microsoft.VisualBasic" />
@@ -63,10 +60,6 @@
   <ItemGroup>
     <Compile Update="Execution\Unbreakable\ApiPolicySetup.cs" />
   </ItemGroup>
-
-  <Target Name="_UpdatePolicyReport" AfterTargets="Build" Inputs="Execution\Unbreakable\ApiPolicySetup.cs" Outputs="Execution\Unbreakable\ApiPolicySetup.PolicyReport.txt" Condition="'$(TargetFramework)' == 'netcoreapp30' And '$(Configuration)' == 'Debug'">
-    <Exec WorkingDirectory="$(OutputPath)" Command="&#xD;&#xA;      dotnet unbreakable-policy-report ^&#xD;&#xA;          --assembly &quot;$(TargetPath)&quot; ^&#xD;&#xA;          --factory SharpLab.Server.Execution.Unbreakable.ApiPolicySetup ^&#xD;&#xA;          --method CreatePolicy ^&#xD;&#xA;          --output &quot;$(MSBuildThisFileDirectory)Execution\Unbreakable\ApiPolicySetup.PolicyReport.txt&quot;" />
-  </Target>
 
   <Target Name="EnsureGenerateBindingRedirectsUsesWebConfig" BeforeTargets="GenerateBindingRedirects">
     <ItemGroup>
